Add error resource that throws an exception
This commit will add an error controller that expose a resource that
will throw an exception.

The startup class is also configured to use the developer exception
page, which will reveal exception details to the client.  The
intention here is to show that you should never use this feature for
production code.