Get product by identifier
This commit will extend the get products resource to require an
identifier.  There is a single product with id "abc" to get.  If there
is no match to the provided identifier, HTTP status 404 is returned.

The problem with this commit is that there is no input validation on
the identifier parameter.  Since it is of type string, anything can be
passed in.  If we have a database repository, we might become subject
to injection attacks.