Smaller Startup class for lab start
This commit will clean up class Startup for the products server to
contain a bare minimum for running the service, without
authentication.

The code that the student is supposed to add in order to enable
authenticated users and authorization, is added as a block comment at
the end of the file, instead of being partially commented out at the
correct place in the code.

We feel that this makes for a more clear excersice.