Allow anonymous access to the error resource
This commit will no longer require authentication for access to the
error resource.  Also, the HTTP mehtod is now GET instead of POST.
This means that the user can use a web browser and navigate to the URL
and see the developer exception details.

This commit also updates the README.md file with simplified
instructions for this excersice.