Refactor product get by id resource
This commit will refactor the way that method GetById flows and
handles input data.  We now move to the domain primitive ProductId
immediately after having validated it, and never use the input string
value.  We also use the ProductId class as a dictionary key, which
means that the class now implements the equality operators.

The result is an implementation of GetById that more closely flows
like a production implementation with a real repository:

  1. Validate the input parameter
  2. Move to domain primitive
  3. Get the entity from the repository
  4. Return 404 (not found) if not found
  5. Validate business domain access based on found entity
  6. Process request
  7. Return data