Update README.md
This commit will update the README.md file to better describe the
steps of this excersice, and explain that the tip of this branch is
the start of the excersice, and you can roll back the latest commit to
see the intended end result of the lab.

The instructions for starting the services are now correct, and the
identity service now support command line arguments for selecting a
port, for completeness.