Update README.md
This commit will update the README.md file to better describe the
steps of this excersice, and explain that the tip of this branch is
the intended end result of the excercise.

The instructions for starting the services are now correct, and the
identity service now support command line arguments for selecting a
port, for completeness.