Refactor code
This commit will refactor the code to make responsibilities a little
more clear, with better separated tests, compared to the previous
commit.  Business logic is split out into a class ProductsService,
which means that the controller class responsibility is only hosting
and input data validation, not business logic.

The instructions for starting the services are now correct, and the
identity service now support command line arguments for selecting a
port, for completeness.