Update product service to use token service
This commit will update class `Startup` in the product service to
verify tokens from the TokenService.  Notice how we replace the use of
the Authority property with hand-coded values for Issuer and Audience.

We also include a JWKs client so that the product service will get the
signing keys from the hosted token service (there is no need for an
installed certificate at the product service.)